requirements.txt
ftfy==6.1.3
numpy==1.24.1
opencv_python==4.8.0.76
pandas==2.0.3
Pillow==10.1.0
regex==2023.8.8
scikit_learn==1.3.1
scipy==1.11.4
timm==0.9.7
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
yacs==0.1.8